Fixing change requests; Moved community agreement to profile and fixed backend user data
