+++ frontend/src/app/about/about.component.html
@@ -25,18 +25,3 @@
     </p>
   </mat-card-content>
 </mat-card>
-
-<mat-card appearance="outlined">
-  <mat-card-title style="margin-left: 10px; margin-top: 10px">
-    Community Agreement</mat-card-title
-  >
-  <mat-card-actions class="token">
-    <button
-      class="agreement-button"
-      mat-stroked-button
-      type="button"
-      (click)="openAgreementDialog()">
-      View Agreement
-    </button>
-  </mat-card-actions>
-</mat-card>
