+++ frontend/src/app/profile/profile-editor/profile-editor.component.html
@@ -97,6 +97,21 @@
 </mat-card>
 
 <mat-card *ngIf="profile.id" appearance="outlined">
+  <mat-card-title style="margin-left: 10px; margin-top: 10px">
+    Community Agreement</mat-card-title
+  >
+  <mat-card-actions class="token">
+    <button
+      class="agreement-button"
+      mat-stroked-button
+      type="button"
+      (click)="openAgreementDialog()">
+      View Agreement
+    </button>
+  </mat-card-actions>
+</mat-card>
+
+<mat-card *ngIf="profile.id" appearance="outlined">
   <mat-card-actions class="token">
     <button
       class="display"
